Fixed save and load error
Save and load button wasn't uploading the new code. It was just getting
it from the DOM.
To recreate (in the old code) put a breakpoint in saveAndLoad(), change
the code in the text area, press save and load, and inspect the variable
strCode. It shouldn't have the new content.